+++ taxel-gui/CHANGELOG.md
@@ -13,6 +13,8 @@
   - Save report at location ("Save as")
   - Delete report from report list
   - Open confirmation PDF after sending the report
+  - Filter income statement based on selected income statement format (e.g.
+    "Gesamtkostenverfahren" or "Umsatzkostenverfahren")
 - changed
   - Add description to "Import values" modal